+++ toolium/__init__.py
@@ -19,4 +19,7 @@ limitations under the License.
 from toolium.driver_wrapper import DriverWrapper
 
 # Get wrapper instance
-toolium_driver = DriverWrapper()
+toolium_wrapper = DriverWrapper(main_driver=True)
+
+# Backwards compatibility
+toolium_driver = toolium_wrapper