+++ toolium/pageelements/input_text_page_element.py
@@ -17,7 +17,6 @@ limitations under the License.
 """
 
 from toolium.pageelements.page_element import PageElement
-from toolium import toolium_driver
 
 
 class InputText(PageElement):
@@ -27,7 +26,7 @@ class InputText(PageElement):
 
     @text.setter
     def text(self, value):
-        if toolium_driver.is_ios_test() and not toolium_driver.is_web_test():
+        if self.driver_wrapper.is_ios_test() and not self.driver_wrapper.is_web_test():
             self.element().set_value(value)
         else:
             self.element().send_keys(value)