+++ toolium/pageobjects/page_object.py
@@ -16,10 +16,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-import unittest
 import logging
+import unittest
 
-from toolium import toolium_driver
+from toolium import toolium_wrapper
 from toolium.pageelements.page_element import PageElement
 from toolium.test_cases import AppiumTestCase
 from toolium.utils import Utils
@@ -27,18 +27,32 @@ from toolium.utils import Utils
 
 class PageObject(unittest.TestCase):
     """
+    :type driver_wrapper: toolium.driver_wrapper.DriverWrapper
     :type driver: selenium.webdriver.remote.webdriver.WebDriver
+    :type config: toolium.config_parser.ExtendedConfigParser
     :type utils: toolium.utils.Utils
     """
 
-    def __init__(self, driver=None):
+    def __init__(self, driver_wrapper=None):
+        """Initialize page object properties and update their page elements
+
+        :param driver_wrapper: driver wrapper instance
+        """
         self.logger = logging.getLogger(__name__)
-        self.driver = driver if driver else toolium_driver.driver
-        self.utils = Utils(self.driver)
-        self.config = toolium_driver.config
+        self.set_driver_wrapper(driver_wrapper)
         self.app_strings = AppiumTestCase.app_strings
         self.init_page_elements()
         self._update_page_elements()
+
+    def set_driver_wrapper(self, driver_wrapper=None):
+        """Initialize driver_wrapper, driver, config and utils
+
+        :param driver_wrapper: driver wrapper instance
+        """
+        self.driver_wrapper = driver_wrapper if driver_wrapper else toolium_wrapper
+        self.set_driver(self.driver_wrapper.driver)
+        self.set_config(self.driver_wrapper.config)
+        self.set_utils(Utils(self.driver))
 
     def set_driver(self, driver):
         """Set Selenium driver"""
@@ -48,6 +62,10 @@ class PageObject(unittest.TestCase):
         """Set utils instance"""
         self.utils = utils
 
+    def set_config(self, config):
+        """Set configuration properties"""
+        self.config = config
+
     def init_page_elements(self):
         """Method to initialize page elements"""
         pass
@@ -56,10 +74,8 @@ class PageObject(unittest.TestCase):
         """Copy driver and utils instances to all page elements of this page object"""
         for element in list(self.__dict__.values()) + list(self.__class__.__dict__.values()):
             if isinstance(element, PageElement):
-                element.set_driver(self.driver)
-                element.set_utils(self.utils)
+                element.set_driver_wrapper(self.driver_wrapper)
             if isinstance(element, PageObject):
-                element.set_driver(self.driver)
-                element.set_utils(self.utils)
+                element.set_driver_wrapper(self.driver_wrapper)
                 # If element is page object, update its page elements
                 element._update_page_elements()