+++ toolium/test_cases.py
@@ -16,15 +16,15 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-import unittest
 import logging
 import sys
+import unittest
 
-from toolium import toolium_driver
-from toolium.utils import Utils
-from toolium.jira import change_all_jira_status
-from toolium.visual_test import VisualTest
+from toolium import toolium_wrapper
 from toolium.config_driver import get_error_message_from_exception
+from toolium.jira import change_all_jira_status
+from toolium.utils import Utils
+from toolium.visual_test import VisualTest
 
 
 class BasicTestCase(unittest.TestCase):
@@ -88,11 +88,11 @@ class BasicTestCase(unittest.TestCase):
     def setUp(self):
         # Configure logger and properties
         if not isinstance(self, SeleniumTestCase):
-            toolium_driver.configure(False, self._config_directory, self._output_directory,
-                                     self._config_properties_filenames, self._config_log_filename,
-                                     self._output_log_filename)
+            toolium_wrapper.configure(False, self._config_directory, self._output_directory,
+                                      self._config_properties_filenames, self._config_log_filename,
+                                      self._output_log_filename)
         # Get config and logger instances
-        self.config = toolium_driver.config
+        self.config = toolium_wrapper.config
         self.logger = logging.getLogger(__name__)
         self.logger.info("Running new test: {0}".format(self.get_subclassmethod_name()))
 
@@ -127,6 +127,7 @@ class SeleniumTestCase(BasicTestCase):
     Attributes:
         driver: webdriver instance
         utils: test utils instance
+        additional_drivers: additional webdriver instances
         remote_video_node: hostname of the remote node if it has enabled a video recorder
 
     :type driver: selenium.webdriver.remote.webdriver.WebDriver
@@ -134,6 +135,7 @@ class SeleniumTestCase(BasicTestCase):
     """
     driver = None
     utils = None
+    additional_drivers = []
     remote_video_node = None
 
     @classmethod
@@ -155,9 +157,12 @@ class SeleniumTestCase(BasicTestCase):
         cls.driver.quit()
         cls.driver = None
         SeleniumTestCase.driver = None
+        for driver in cls.additional_drivers:
+            driver.quit()
+        cls.additional_drivers = []
 
         # Download saved video if video is enabled or if test fails
-        if cls.remote_video_node and (toolium_driver.config.getboolean_optional('Server', 'video_enabled') or
+        if cls.remote_video_node and (toolium_wrapper.config.getboolean_optional('Server', 'video_enabled') or
                                           not test_passed):
             video_name = video_name if test_passed else 'error_{}'.format(video_name)
             cls.utils.download_remote_video(cls.remote_video_node, session_id, video_name)
@@ -165,19 +170,16 @@ class SeleniumTestCase(BasicTestCase):
     def setUp(self):
         # Create driver
         if not SeleniumTestCase.driver:
-            toolium_driver.configure(True, self._config_directory, self._output_directory,
-                                     self._config_properties_filenames, self._config_log_filename,
-                                     self._output_log_filename)
-            SeleniumTestCase.driver = toolium_driver.connect()
+            toolium_wrapper.configure(True, self._config_directory, self._output_directory,
+                                      self._config_properties_filenames, self._config_log_filename,
+                                      self._output_log_filename)
+            SeleniumTestCase.driver = toolium_wrapper.connect()
             SeleniumTestCase.utils = Utils(SeleniumTestCase.driver)
             SeleniumTestCase.remote_video_node = SeleniumTestCase.utils.get_remote_video_node()
         # Get common configuration of reusing driver
-        self.reuse_driver = toolium_driver.config.getboolean_optional('Common', 'reuse_driver')
+        self.reuse_driver = toolium_wrapper.config.getboolean_optional('Common', 'reuse_driver')
         # Set implicitly wait
         self.utils.set_implicit_wait()
-        # Maximize browser
-        if toolium_driver.is_maximizable():
-            SeleniumTestCase.driver.maximize_window()
         # Call BasicTestCase setUp
         super(SeleniumTestCase, self).setUp()
 
@@ -189,12 +191,16 @@ class SeleniumTestCase(BasicTestCase):
         test_name = self.get_subclassmethod_name().replace('.', '_')
         if not self._test_passed:
             self.utils.capture_screenshot(test_name)
+            driver_index = 1
+            for driver in self.additional_drivers:
+                driver_index += 1
+                Utils(driver).capture_screenshot('{}_driver{}'.format(test_name, driver_index))
 
         # Stop driver
         if not self.reuse_driver:
             SeleniumTestCase._finalize_driver(test_name, self._test_passed)
 
-    def assertScreenshot(self, element_or_selector, filename, threshold=0, exclude_elements=[]):
+    def assertScreenshot(self, element_or_selector, filename, threshold=0, exclude_elements=[], driver_wrapper=None):
         """Assert that a screenshot of an element is the same as a screenshot on disk, within a given threshold.
 
         :param element_or_selector: either a CSS/XPATH selector as a string or a WebElement object.
@@ -203,20 +209,23 @@ class SeleniumTestCase(BasicTestCase):
         :param threshold: the threshold for triggering a test failure
         :param exclude_elements: list of CSS/XPATH selectors as a string or WebElement objects that must be excluded
                                  from the assertion.
+        :param driver_wrapper: driver wrapper instance
         """
         file_suffix = self.get_method_name()
-        VisualTest().assertScreenshot(element_or_selector, filename, file_suffix, threshold, exclude_elements)
+        VisualTest(driver_wrapper).assertScreenshot(element_or_selector, filename, file_suffix, threshold,
+                                                    exclude_elements)
 
-    def assertFullScreenshot(self, filename, threshold=0, exclude_elements=[]):
+    def assertFullScreenshot(self, filename, threshold=0, exclude_elements=[], driver_wrapper=None):
         """Assert that a driver screenshot is the same as a screenshot on disk, within a given threshold.
 
         :param filename: the filename for the screenshot, which will be appended with ``.png``
         :param threshold: the threshold for triggering a test failure
         :param exclude_elements: list of CSS/XPATH selectors as a string or WebElement objects that must be excluded
                                  from the assertion.
+        :param driver_wrapper: driver wrapper instance
         """
         file_suffix = self.get_method_name()
-        VisualTest().assertScreenshot(None, filename, file_suffix, threshold, exclude_elements)
+        VisualTest(driver_wrapper).assertScreenshot(None, filename, file_suffix, threshold, exclude_elements)
 
 
 class AppiumTestCase(SeleniumTestCase):
@@ -239,7 +248,7 @@ class AppiumTestCase(SeleniumTestCase):
 
     def setUp(self):
         super(AppiumTestCase, self).setUp()
-        if AppiumTestCase.app_strings is None and not toolium_driver.is_web_test():
+        if AppiumTestCase.app_strings is None and not toolium_wrapper.is_web_test():
             AppiumTestCase.app_strings = self.driver.app_strings()
 
     def tearDown(self):